+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 8.6
 -----
+
+- [Dynamic Type] Update UpNext [#3918](https://github.com/Automattic/pocket-casts-ios/pull/3918)
 - [TestFlight only] Add a new VoiceBoost solution designed to maintain more consistent audio levels. [#3915](https://github.com/Automattic/pocket-casts-ios/pull/3915)
 - Dynamic Type: Update Podcasts list [#3914](https://github.com/Automattic/pocket-casts-ios/pull/3914)
 - Dynamic Type: Update Podcast detail [#3892](https://github.com/Automattic/pocket-casts-ios/pull/3892)